+++ type-definitions/ts-tests/map.ts
@@ -492,7 +492,6 @@ import { Map, List } from 'immutable';
 {
   // #toJS
 
-  // should be `{ [x: string]: number; [x: number]: number; [x: symbol]: number; }` but there is an issue with circular references
-  // $ExpectType { [x: string]: unknown; [x: number]: unknown; [x: symbol]: unknown; }
+  // $ExpectType { [x: string]: number; [x: number]: number; [x: symbol]: number; }
   Map<number, number>().toJS();
 }